+++ app/product/biz/model/base.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Base struct {
-	ID        int `gorm:"primarykey"`
+	ID        int64 `gorm:"primarykey"`
 	CreatedAt time.Time
 	UpdatedAt time.Time
 	DeletedAt *time.Time
